refactored about DI sources
